Moved several components into their own folder
Also fixed some old unused exports.
