+++ src/components/index.ts
@@ -1,5 +1,2 @@
-export { Header } from './Header';
-export { Post } from './Post';
-export { Posts } from './Posts';
-export { DiscordButton } from './DiscordButton';
-export { Test } from './Test';
+export { Home } from '@components/Home';
+export { Test } from '@components/Test';
