+++ src/components/wrapper/index.ts
@@ -1,0 +1,3 @@
+export { Logo } from '@components/wrapper/Logo';
+export { Header } from '@components/wrapper/Header';
+export { Footer } from '@components/wrapper/Footer';
