+++ src/layouts/Wrapper.astro
@@ -1,7 +1,6 @@
 ---
 import 'normalize.css';
-import Header from '@components/Header';
-import Footer from '@components/Footer';
+import { Header, Footer } from "@components/wrapper/index";
 
 export interface Props {
 	title: string;
